requirements.txt
streamlit
speechrecognition
pyttsx3
shapely<2.0.0
google-cloud-aiplatform>=1.27.0
google-cloud-texttospeech
gTTS
pydub